+++ test/Middleware/ImplicitHeadMiddlewareTest.php
@@ -1,0 +1,168 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-router for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-router/blob/master/LICENSE.md New BSD License
+ */
+
+namespace ZendTest\Expressive\Router\Middleware;
+
+use Fig\Http\Message\RequestMethodInterface as RequestMethod;
+use PHPUnit\Framework\TestCase;
+use Prophecy\Argument;
+use Prophecy\Prophecy\ObjectProphecy;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Message\StreamInterface;
+use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as RequestHandlerInterface;
+use Zend\Expressive\Router\Middleware\ImplicitHeadMiddleware;
+use Zend\Expressive\Router\Route;
+use Zend\Expressive\Router\RouteResult;
+use Zend\Expressive\Router\RouterInterface;
+
+use const Webimpress\HttpMiddlewareCompatibility\HANDLER_METHOD;
+
+class ImplicitHeadMiddlewareTest extends TestCase
+{
+    /** @var ImplicitHeadMiddleware */
+    private $middleware;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $response;
+
+    /** @var StreamInterface|ObjectProphecy */
+    private $stream;
+
+    public function setUp()
+    {
+        $this->response = $this->prophesize(ResponseInterface::class);
+
+        $this->stream = $this->prophesize(StreamInterface::class);
+        $streamFactory = function () {
+            return $this->stream->reveal();
+        };
+
+        $this->middleware = new ImplicitHeadMiddleware($this->response->reveal(), $streamFactory);
+    }
+
+    public function testReturnsResultOfHandlerOnNonHeadRequests()
+    {
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_GET);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->{HANDLER_METHOD}($request->reveal())->will([$this->response, 'reveal']);
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($this->response->reveal(), $result);
+    }
+
+    public function testReturnsResultOfHandlerWhenNoRouteResultPresentInRequest()
+    {
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
+        $request->getAttribute(RouteResult::class)->willReturn(null);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->{HANDLER_METHOD}($request->reveal())->will([$this->response, 'reveal']);
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($this->response->reveal(), $result);
+    }
+
+    public function testReturnsResultOfHandlerWhenRouteResultDoesNotContainAMatchedRoute()
+    {
+        $result = $this->prophesize(RouteResult::class);
+        $result->getMatchedRoute()->willReturn(null);
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
+        $request->getAttribute(RouteResult::class)->will([$result, 'reveal']);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->{HANDLER_METHOD}($request->reveal())->will([$this->response, 'reveal']);
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($this->response->reveal(), $result);
+    }
+
+    public function testReturnsResultOfHandlerWhenRouteResultContainsAMatchedRouteWithExplicitHeadSupport()
+    {
+        $route = $this->prophesize(Route::class);
+        $route->implicitHead()->willReturn(false);
+
+        $result = $this->prophesize(RouteResult::class);
+        $result->getMatchedRoute()->will([$route, 'reveal']);
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
+        $request->getAttribute(RouteResult::class)->will([$result, 'reveal']);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->{HANDLER_METHOD}($request->reveal())->will([$this->response, 'reveal']);
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($this->response->reveal(), $result);
+    }
+
+    public function testReturnsCannedResponseWhenRouteDoesNotExplicitlySupportHeadAndDoesNotSupportGet()
+    {
+        $route = $this->prophesize(Route::class);
+        $route->implicitHead()->willReturn(true);
+        $route->allowsMethod(RequestMethod::METHOD_GET)->willReturn(false);
+
+        $result = $this->prophesize(RouteResult::class);
+        $result->getMatchedRoute()->will([$route, 'reveal']);
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
+        $request->getAttribute(RouteResult::class)->will([$result, 'reveal']);
+        $request->withMethod(RequestMethod::METHOD_GET)->will([$request, 'reveal']);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+        $handler->{HANDLER_METHOD}(Argument::any())->shouldNotBeCalled();
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($this->response->reveal(), $result);
+    }
+
+    public function testInvokesHandlerWhenRouteImplicitlySupportsHeadAndSupportsGet()
+    {
+        $headRoute = $this->prophesize(Route::class);
+        $headRoute->implicitHead()->willReturn(true);
+        $headRoute->allowsMethod(RequestMethod::METHOD_GET)->willReturn(true);
+
+        $headResult = $this->prophesize(RouteResult::class);
+        $headResult->getMatchedRoute()->will([$headRoute, 'reveal']);
+
+        $handler = $this->prophesize(RequestHandlerInterface::class);
+
+        $response = $this->prophesize(ResponseInterface::class);
+        $response->withBody($this->stream->reveal())->will([$response, 'reveal']);
+
+        $request = $this->prophesize(ServerRequestInterface::class);
+        $request->getMethod()->willReturn(RequestMethod::METHOD_HEAD);
+        $request->getAttribute(RouteResult::class)->will([$headResult, 'reveal']);
+        $request->withMethod(RequestMethod::METHOD_GET)->will([$request, 'reveal']);
+        $request
+            ->withAttribute(
+                ImplicitHeadMiddleware::FORWARDED_HTTP_METHOD_ATTRIBUTE,
+                RequestMethod::METHOD_HEAD
+            )
+            ->will(function () use ($request, $response, $handler) {
+                $handler
+                    ->{HANDLER_METHOD}(Argument::that([$request, 'reveal']))
+                    ->will([$response, 'reveal']);
+                return $request->reveal();
+            });
+
+        $result = $this->middleware->process($request->reveal(), $handler->reveal());
+
+        $this->assertSame($response->reveal(), $result);
+    }
+}